似乎无意义的测试。。。

+++ 考研/数据结构/Sort.cpp
@@ -57,7 +57,7 @@ void DisplayVector(vector<ElemType> &vec, size_t rowLength)
     }
 }
 
-ssize_t TimeConsumed(void (*func)(vector<ElemType> &), vector<ElemType> &vec)
+int64_t TimeConsumed(void (*func)(vector<ElemType> &), vector<ElemType> &vec)
 {
     auto start = chrono::high_resolution_clock::now();
     func(vec);
